marketplaceyoasi/marketplace/serializer.py
from rest_framework import serializers
from .models import Product, Seller


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ["name", "description", "price", "seller"]


class SellerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Seller
        fields = ["id", "name", "cnpj", "phone"]